+++ packages/blocks/core/js/wordpress/navigation-link/test/functionality.blocks-navigation-link.e2e.cy.js
@@ -7,6 +7,7 @@ import {
 	appendBlocks,
 	openInserter,
 	setInnerBlock,
+	setParentBlock,
 	redirectToFrontPage,
 } from '@blockera/dev-cypress/js/helpers';
 
@@ -122,7 +123,20 @@ describe(
 				});
 
 			//
-			// 2. Assert inner blocks selectors in front end
+			// 3. Check settings tab
+			//
+			setParentBlock();
+			cy.getByDataTest('settings-tab').click();
+
+			cy.get('.block-editor-block-inspector').within(() => {
+				cy.get('.components-tools-panel-header')
+					.contains('Settings')
+					.scrollIntoView()
+					.should('be.visible');
+			});
+
+			//
+			// 4. Assert inner blocks selectors in front end
 			//
 			savePage();
 			redirectToFrontPage();